fix(material): D-414 领料单明细数量支持小数（INT→DECIMAL 12,4）
领料单 t_material_picking_item.quantity 是小数全链路最后一个取整点，
承接 D-410/D-414 统一为 DECIMAL(12,4)，避免按 1.32 领料落库 1 导致
出库流水/库存/回冲锁定量对不上。同时入库流程不再自动生成物料对账，
改由用户「入库到仓库/直接使用」选择决定账目走向，消除重复对账记录。

+++ backend/src/main/java/com/fashion/supplychain/intelligence/agent/tool/MaterialPickingTool.java
@@ -14,6 +14,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.context.annotation.Lazy;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -122,7 +123,8 @@ public class MaterialPickingTool extends AbstractAgentTool {
                 if (quantity != null) {
                     MaterialPickingItem item = new MaterialPickingItem();
                     item.setMaterialName(materialName);
-                    item.setQuantity(quantity);
+                    // D-414：领料数量字段已是 BigDecimal
+                    item.setQuantity(BigDecimal.valueOf(quantity));
                     item.setUnit(unit != null ? unit : "米");
                     item.setTenantId(tenantId);
                     items.add(item);
